Added SQL Inserts

+++ src/main/java/bchr/stdb/dao/AccountDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AccountDao extends JpaRepository<Account,Integer> {
+public interface AccountDao extends JpaRepository<Account,Long> {
     boolean existsByCompName(String compName);
 
     Account findByCompName(String compName);
